Harmonised folder structure and file naming.
Removed sub-directory detector from microqrcode directory.
Renamed all micro QR code files to start with MQR.
