+++ core/src/microqrcode/MQRDimension.cpp
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#include "Dimension.h"
+#include "MQRDimension.h"
 
 using ZXing::MicroQRCode::Dimension;
