+++ core/src/microqrcode/MQRFinderPattern.h
