+++ core/src/microqrcode/MQRNotFoundException.h
