+++ core/src/microqrcode/MQRReaderException.h
